feat: myScreen 로직 변경
