+++ app/src/main/java/org/sopt/at/my/MyActivity.kt
@@ -1,0 +1,49 @@
+package org.sopt.at.my
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.lifecycle.ViewModelProvider
+import org.sopt.at.login.ui.signin.SignInActivity
+import org.sopt.at.login.util.LoginRepository
+import org.sopt.at.login.util.LoginViewModelFactory
+import org.sopt.at.ui.theme.ATSOPTANDROIDTheme
+
+class MyActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        val repository = LoginRepository(getSharedPreferences("prefs", MODE_PRIVATE))
+        val viewModelFactory = LoginViewModelFactory(repository)
+        val viewModel = ViewModelProvider(this, viewModelFactory).get(MyViewModel::class.java)
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+
+        setContent {
+            ATSOPTANDROIDTheme {
+                MyScreen(viewModel)
+            }
+        }
+    }
+}
