+++ app/src/main/java/org/sopt/at/my/MyScreen.kt
@@ -1,12 +1,10 @@
-package org.sopt.at.login.ui
+package org.sopt.at.my
 
+import android.app.Activity
 import android.content.Intent
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
@@ -18,46 +16,34 @@ import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import org.sopt.at.util.SharedPreferencesManager
-import org.sopt.at.ui.theme.ATSOPTANDROIDTheme
+import org.sopt.at.login.ui.signin.SignInActivity
 
-class MyActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-
-        setContent {
-            ATSOPTANDROIDTheme {
-                MyScreen()
-            }
-        }
-    }
-}
-
-@Preview(showBackground = true)
 @Composable
-fun MyScreen(){
+fun MyScreen(viewModel: MyViewModel) {
+    val userId by viewModel.userId.collectAsState()
     val context = LocalContext.current
-    val (userId, userPassword) = SharedPreferencesManager.getUser(context)
+    LaunchedEffect(Unit) {
+        viewModel.getUserInfo("")
+    }
     Scaffold(
         modifier = Modifier.fillMaxSize().background(color = Color.Black).padding(horizontal = 15.dp),
         containerColor = Color.Black,
         bottomBar ={
             Button(
                 onClick = {
-                    SharedPreferencesManager.logOut(context)
-                    val intent = Intent(context, SignInActivity::class.java).apply {
-                        flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
-                    }
-                    context.startActivity(intent)
+                   viewModel.logOut()
+                    context.startActivity(Intent(context, SignInActivity::class.java))
+                    (context as? Activity)?.finish()
 
                 },
                 modifier = Modifier.fillMaxWidth().padding(vertical = 40.dp)
@@ -67,7 +53,7 @@ fun MyScreen(){
                     containerColor = Color.Black,
                     contentColor = Color.White),
                 contentPadding = PaddingValues(vertical = 15.dp)
-                ) {
+            ) {
                 Text("로그아웃",
                     fontSize = 17.sp,
                     fontWeight = FontWeight.Bold,
